Login router fix #42
