+++ src/main.js
@@ -1,5 +1,6 @@
 import Vue from 'vue'
 import App from './App.vue'
+import router from './route/index.js'
 import './registerServiceWorker'
 
 Vue.config.productionTip = false
@@ -26,6 +27,8 @@ const firebaseConfig = {
 initializeApp(firebaseConfig);
 // const analytics = getAnalytics(app);
 
-new Vue({
-  render: h => h(App),
+new Vue({  
+  router,
+  render: h => h(App)
 }).$mount('#app')
+
